Day 18 working. Part1 implementation was correct, but there was a stupid regex error. It's fixed now.

+++ src/main/kotlin/year23/day18/Coordinate.kt
@@ -1,6 +1,6 @@
 package year23.day18
 
 data class Coordinate(
-    val row: Int,
-    val col: Int
+    val row: Long,
+    val col: Long
 )
